requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pdfplumber==0.10.2
python-docx==1.1.0
python-multipart==0.0.9
gunicorn==21.2.0
tensorflow-cpu==2.12.0
tensorflow-hub==0.14.0
